+++ Geometry/GeometryTest/TestDeckedSlabBeam.cpp
@@ -624,7 +624,7 @@ void CTestDeckedSlabBeam::TestIXYPosition()
    TRY_TEST( cPoints,11);
 
    coll->get__Enum(&Enum);
-   Enum->Next(cPoints,&points[0],&fetched);
+   Enum->Next((ULONG)cPoints,&points[0],&fetched);
    TRY_TEST( fetched, cPoints );
 
    i = 0;
